+++ modules/admin/src/filters/SmallCrop.php
@@ -9,7 +9,7 @@ namespace admin\filters;
  */
 class SmallCrop extends \admin\base\Filter
 {
-    public function identifier()
+    public static function identifier()
     {
         return 'small-crop';
     }